Move all App details to app mod
