+++ payjoin-client/src/main.rs
@@ -1,23 +1,8 @@
-use std::collections::HashMap;
-use std::convert::TryFrom;
-use std::str::FromStr;
-
-use anyhow::{anyhow, Context, Result};
-use bitcoincore_rpc::bitcoin::Amount;
-use bitcoincore_rpc::RpcApi;
+use anyhow::{Context, Result};
 use clap::{arg, Arg, ArgMatches, Command};
-use payjoin::bitcoin::util::psbt::PartiallySignedTransaction as Psbt;
-use payjoin::receive::{Error, PayjoinProposal};
-use payjoin::{PjUriExt, UriExt};
-use rouille::{Request, Response};
 
-mod app_config;
-use app_config::AppConfig;
-
-struct App {
-    config: AppConfig,
-    bitcoind: bitcoincore_rpc::Client,
-}
+mod app;
+use app::{App, AppConfig};
 
 fn main() -> Result<()> {
     env_logger::init();
@@ -40,337 +25,6 @@ fn main() -> Result<()> {
     }
 
     Ok(())
-}
-
-impl App {
-    fn new(config: AppConfig) -> Result<Self> {
-        let bitcoind = match &config.bitcoind_cookie {
-            Some(cookie) => bitcoincore_rpc::Client::new(
-                &config.bitcoind_rpchost,
-                bitcoincore_rpc::Auth::CookieFile(cookie.into()),
-            ),
-            None => bitcoincore_rpc::Client::new(
-                &config.bitcoind_rpchost,
-                bitcoincore_rpc::Auth::UserPass(
-                    config.bitcoind_rpcuser.clone(),
-                    config.bitcoind_rpcpass.clone(),
-                ),
-            ),
-        }
-        .context("Failed to connect to bitcoind")?;
-        Ok(Self { config, bitcoind })
-    }
-
-    fn send_payjoin(&self, bip21: &str) -> Result<()> {
-        let link = payjoin::Uri::try_from(bip21)
-            .map_err(|e| anyhow!("Failed to create URI from BIP21: {}", e))?;
-
-        let link = link
-            .check_pj_supported()
-            .map_err(|e| anyhow!("The provided URI doesn't support payjoin (BIP78): {}", e))?;
-
-        let amount = link
-            .amount
-            .ok_or_else(|| anyhow!("please specify the amount in the Uri"))
-            .map(|amt| Amount::from_sat(amt.to_sat()))?;
-        let mut outputs = HashMap::with_capacity(1);
-        outputs.insert(link.address.to_string(), amount);
-
-        let options = bitcoincore_rpc::json::WalletCreateFundedPsbtOptions {
-            lock_unspent: Some(true),
-            fee_rate: Some(Amount::from_sat(2000)),
-            ..Default::default()
-        };
-        let psbt = self
-            .bitcoind
-            .wallet_create_funded_psbt(
-                &[], // inputs
-                &outputs,
-                None, // locktime
-                Some(options),
-                None,
-            )
-            .context("Failed to create PSBT")?
-            .psbt;
-        let psbt = self
-            .bitcoind
-            .wallet_process_psbt(&psbt, None, None, None)
-            .with_context(|| "Failed to process PSBT")?
-            .psbt;
-        let psbt = load_psbt_from_base64(psbt.as_bytes())
-            .with_context(|| "Failed to load PSBT from base64")?;
-        log::debug!("Original psbt: {:#?}", psbt);
-        let pj_params = payjoin::send::Configuration::with_fee_contribution(
-            payjoin::bitcoin::Amount::from_sat(10000),
-            None,
-        );
-        let (req, ctx) = link
-            .create_pj_request(psbt, pj_params)
-            .with_context(|| "Failed to create payjoin request")?;
-        let client = reqwest::blocking::Client::builder()
-            .danger_accept_invalid_certs(self.config.danger_accept_invalid_certs)
-            .build()
-            .with_context(|| "Failed to build reqwest http client")?;
-        let response = client
-            .post(req.url)
-            .body(req.body)
-            .header("Content-Type", "text/plain")
-            .send()
-            .with_context(|| "HTTP request failed")?;
-        // TODO display well-known errors and log::debug the rest
-        let psbt = ctx.process_response(response).with_context(|| "Failed to process response")?;
-        log::debug!("Proposed psbt: {:#?}", psbt);
-        let psbt = self
-            .bitcoind
-            .wallet_process_psbt(&serialize_psbt(&psbt), None, None, None)
-            .with_context(|| "Failed to process PSBT")?
-            .psbt;
-        let tx = self
-            .bitcoind
-            .finalize_psbt(&psbt, Some(true))
-            .with_context(|| "Failed to finalize PSBT")?
-            .hex
-            .ok_or_else(|| anyhow!("Incomplete PSBT"))?;
-        let txid = self
-            .bitcoind
-            .send_raw_transaction(&tx)
-            .with_context(|| "Failed to send raw transaction")?;
-        log::info!("Transaction sent: {}", txid);
-        Ok(())
-    }
-
-    fn receive_payjoin(self, amount_arg: &str) -> Result<()> {
-        use payjoin::Uri;
-
-        let pj_receiver_address = self.bitcoind.get_new_address(None, None)?;
-        let amount = Amount::from_sat(amount_arg.parse()?);
-        let pj_uri_string = format!(
-            "{}?amount={}&pj={}",
-            pj_receiver_address.to_qr_uri(),
-            amount.to_btc(),
-            self.config.pj_endpoint
-        );
-        let pj_uri = Uri::from_str(&pj_uri_string)
-            .map_err(|e| anyhow!("Constructed a bad URI string from args: {}", e))?;
-        let _pj_uri = pj_uri
-            .check_pj_supported()
-            .map_err(|e| anyhow!("Constructed URI does not support payjoin: {}", e))?;
-
-        println!("Awaiting payjoin at BIP 21 Payjoin Uri:");
-        println!("{}", pj_uri_string);
-
-        rouille::start_server(self.config.pj_host.clone(), move |req| self.handle_web_request(req));
-    }
-
-    fn handle_web_request(&self, req: &Request) -> Response {
-        self.handle_payjoin_request(req)
-            .map_err(|e| match e {
-                Error::BadRequest(e) => {
-                    log::error!("Error handling request: {}", e);
-                    Response::text(e.to_string()).with_status_code(400)
-                }
-                e => {
-                    log::error!("Error handling request: {}", e);
-                    Response::text(e.to_string()).with_status_code(500)
-                }
-            })
-            .unwrap_or_else(|err_resp| err_resp)
-    }
-
-    fn handle_payjoin_request(&self, req: &Request) -> Result<Response, Error> {
-        use bitcoin::hashes::hex::ToHex;
-
-        let headers = Headers(req.headers());
-        let proposal = payjoin::receive::UncheckedProposal::from_request(
-            req.data().context("Failed to read request body").map_err(|e| {
-                log::warn!("Failed to read request body: {}", e);
-                Error::Server(e.into())
-            })?,
-            req.raw_query_string(),
-            headers,
-        )?;
-
-        // in a payment processor where the sender could go offline, this is where you schedule to broadcast the original_tx
-        let _to_broadcast_in_failure_case = proposal.get_transaction_to_schedule_broadcast();
-
-        // The network is used for checks later
-        let network = match self
-            .bitcoind
-            .get_blockchain_info()
-            .map_err(|e| Error::Server(e.into()))?
-            .chain
-            .as_str()
-        {
-            "main" => bitcoin::Network::Bitcoin,
-            "test" => bitcoin::Network::Testnet,
-            "regtest" => bitcoin::Network::Regtest,
-            _ => return Err(Error::Server(anyhow!("Unknown network").into())),
-        };
-
-        // Receive Check 1: Can Broadcast
-        let proposal = proposal.check_can_broadcast(|tx| {
-            let raw_tx = bitcoin::consensus::encode::serialize(&tx).to_hex();
-            let mempool_results = self
-                .bitcoind
-                .test_mempool_accept(&[raw_tx])
-                .map_err(|e| Error::Server(e.into()))?;
-            match mempool_results.first() {
-                Some(result) => Ok(result.allowed),
-                None => Err(Error::Server(
-                    anyhow!("No mempool results returned on broadcast check").into(),
-                )),
-            }
-        })?;
-        log::trace!("check1");
-
-        // Receive Check 2: receiver can't sign for proposal inputs
-        let proposal = proposal.check_inputs_not_owned(|input| {
-            if let Ok(address) = bitcoin::Address::from_script(input, network) {
-                self.bitcoind
-                    .get_address_info(&address)
-                    .map(|info| info.is_mine.unwrap_or(false))
-                    .map_err(|e| Error::Server(e.into()))
-            } else {
-                Ok(false)
-            }
-        })?;
-        log::trace!("check2");
-        // Receive Check 3: receiver can't sign for proposal inputs
-        let proposal = proposal.check_no_mixed_input_scripts()?;
-        log::trace!("check3");
-
-        // Receive Check 4: have we seen this input before? More of a check for non-interactive i.e. payment processor receivers.
-        let payjoin = proposal.check_no_inputs_seen_before(|input| {
-            if let Ok(address) = bitcoin::Address::from_script(input, network) {
-                let address_info = self.bitcoind.get_address_info(&address).map_err(|e| Error::Server(e.into()))?;
-                if address_info.labels.contains(
-                    &bitcoincore_rpc::bitcoincore_rpc_json::GetAddressInfoResultLabel::Simple(
-                        "input_seen_before".into(),
-                    ),
-                ) {
-                    log::info!(
-                        "Request contains an input we've seen before. Preventing possible probing attack: {}",
-                        address
-                    );
-                    Ok(true)
-                } else {
-                    self.bitcoind
-                        .import_address(&address, Some("input_seen_before"), Some(false)).map_err(|e| Error::Server(e.into()))?;
-                    Ok(false)
-                }
-            } else {
-                Ok(false)
-            }
-        })?;
-        log::trace!("check4");
-
-        let mut payjoin = payjoin.identify_receiver_outputs(|output_script| {
-            if let Ok(address) = bitcoin::Address::from_script(output_script, network) {
-                self.bitcoind
-                    .get_address_info(&address)
-                    .map(|info| info.is_mine.unwrap_or(false))
-                    .map_err(|e| Error::Server(e.into()))
-            } else {
-                Ok(false)
-            }
-        })?;
-
-        if !self.config.sub_only {
-            // Select receiver payjoin inputs.
-            _ = try_contributing_inputs(&mut payjoin, &self.bitcoind)
-                .map_err(|e| log::warn!("Failed to contribute inputs: {}", e));
-        }
-
-        let receiver_substitute_address =
-            self.bitcoind.get_new_address(None, None).map_err(|e| Error::Server(e.into()))?;
-        payjoin.substitute_output_address(receiver_substitute_address);
-
-        let payjoin_proposal_psbt = payjoin.apply_fee(Some(1))?;
-
-        log::debug!("Extracted PSBT: {:#?}", payjoin_proposal_psbt);
-        // Sign payjoin psbt
-        let payjoin_base64_string =
-            base64::encode(bitcoin::consensus::serialize(&payjoin_proposal_psbt));
-        // `wallet_process_psbt` adds available utxo data and finalizes
-        let payjoin_proposal_psbt = self
-            .bitcoind
-            .wallet_process_psbt(&payjoin_base64_string, None, None, Some(false))
-            .map_err(|e| Error::Server(e.into()))?
-            .psbt;
-        let payjoin_proposal_psbt = load_psbt_from_base64(payjoin_proposal_psbt.as_bytes())
-            .context("Failed to parse PSBT")
-            .map_err(|e| Error::Server(e.into()))?;
-        let payjoin_proposal_psbt = payjoin.prepare_psbt(payjoin_proposal_psbt)?;
-        log::debug!("Receiver's PayJoin proposal PSBT Rsponse: {:#?}", payjoin_proposal_psbt);
-
-        let payload = base64::encode(bitcoin::consensus::serialize(&payjoin_proposal_psbt));
-        log::info!("successful response");
-        Ok(Response::text(payload))
-    }
-}
-
-fn try_contributing_inputs(
-    payjoin: &mut PayjoinProposal,
-    bitcoind: &bitcoincore_rpc::Client,
-) -> Result<()> {
-    use bitcoin::OutPoint;
-
-    let available_inputs = bitcoind
-        .list_unspent(None, None, None, None, None)
-        .context("Failed to list unspent from bitcoind")?;
-    let candidate_inputs: HashMap<Amount, OutPoint> = available_inputs
-        .iter()
-        .map(|i| (i.amount, OutPoint { txid: i.txid, vout: i.vout }))
-        .collect();
-
-    let selected_outpoint = payjoin.try_preserving_privacy(candidate_inputs).expect("gg");
-    let selected_utxo = available_inputs
-        .iter()
-        .find(|i| i.txid == selected_outpoint.txid && i.vout == selected_outpoint.vout)
-        .context("This shouldn't happen. Failed to retrieve the privacy preserving utxo from those we provided to the seclector.")?;
-    log::debug!("selected utxo: {:#?}", selected_utxo);
-
-    //  calculate receiver payjoin outputs given receiver payjoin inputs and original_psbt,
-    let txo_to_contribute = bitcoin::TxOut {
-        value: selected_utxo.amount.to_sat(),
-        script_pubkey: selected_utxo.script_pub_key.clone(),
-    };
-    let outpoint_to_contribute =
-        bitcoin::OutPoint { txid: selected_utxo.txid, vout: selected_utxo.vout };
-    payjoin.contribute_witness_input(txo_to_contribute, outpoint_to_contribute);
-    Ok(())
-}
-
-struct Headers<'a>(rouille::HeadersIter<'a>);
-impl payjoin::receive::Headers for Headers<'_> {
-    fn get_header(&self, key: &str) -> Option<&str> {
-        let mut copy = self.0.clone(); // lol
-        copy.find(|(k, _)| k.eq_ignore_ascii_case(key)).map(|(_, v)| v)
-    }
-}
-
-fn load_psbt_from_base64(
-    mut input: impl std::io::Read,
-) -> Result<Psbt, payjoin::bitcoin::consensus::encode::Error> {
-    use payjoin::bitcoin::consensus::Decodable;
-
-    let mut reader = base64::read::DecoderReader::new(
-        &mut input,
-        base64::Config::new(base64::CharacterSet::Standard, true),
-    );
-    Psbt::consensus_decode(&mut reader)
-}
-
-fn serialize_psbt(psbt: &Psbt) -> String {
-    use payjoin::bitcoin::consensus::Encodable;
-
-    let mut encoder = base64::write::EncoderWriter::new(Vec::new(), base64::STANDARD);
-    psbt.consensus_encode(&mut encoder)
-        .expect("Vec doesn't return errors in its write implementation");
-    String::from_utf8(
-        encoder.finish().expect("Vec doesn't return errors in its write implementation"),
-    )
-    .unwrap()
 }
 
 fn cli() -> ArgMatches {
